feat(DTFS2-8431-8432): split up amendment payload functions
