+++ trade-finance-manager-api/server/v1/mappings/apim-gift-payloads/amend-facility/get-amount-difference/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Calculates the absolute difference between two amounts.
+ * @param previousAmount - The previous amount.
+ * @param newAmount - The new amount.
+ * @returns The absolute difference between the previous and new amounts.
+ */
+export const getAmountDifference = (previousAmount: number, newAmount: number): number => Math.abs(previousAmount - newAmount);
